% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mooplot-package.R
\docType{package}
\name{mooplot-package}
\alias{mooplot}
\alias{mooplot-package}
\title{mooplot: Graphical Visualizations for Multi-Objective Optimization}
\description{
Visualization of multi-dimensional data arising in multi-objective optimization, including plots of the empirical attainment function (EAF) and symmetric Vorob'ev expectation and deviation.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://multi-objective.github.io/mooplot/r/}
  \item \url{https://github.com/multi-objective/mooplot/}
  \item Report bugs at \url{https://github.com/multi-objective/mooplot/issues/}
}

}
\author{
\strong{Maintainer}: Manuel López-Ibáñez \email{manuel.lopez-ibanez@manchester.ac.uk} (\href{https://orcid.org/0000-0001-9974-1295}{ORCID})

}
\keyword{internal}
